Add a NumPy build extension command class
Ensures that we can still add NumPy's headers to the extensions we plan
to build even though NumPy has not been installed via setup requirements
on the first pass of `setup.py`. Does this by overriding the `build_ext`
command class to add in the NumPy headers as part of option
finalization. Should make it easier for anyone trying to install this
package without NumPy already installed.