Bump NumPy requirement to 1.11.3 for Windows
As there are no Windows builds of NumPy for Python 2.7, 3.4-3.6 for any
NumPy version earlier than 1.11.3, having a NumPy version requirement of
1.8.2 is unattainable on Windows. So bump the NumPy version requirement
to a version of NumPy where there are builds for all supported Python
versions and OSes. The lowest major minor version meeting this
requirements is NumPy 1.11.3. Hence the new NumPy requirement is set to
1.11.3.