Implement a minmax function in C via Cython
Takes an array of `real` values as input and 2 element array of reals
for writing the output to. Iterates over the array storing anything
larger than the max value or anything smaller than the minimum value.
Everything else inbetween is simply passed over. This is done until the
input array is exhausted. For efficiency, the input array and output
arrays are both memoryviews, which allows the GIL to be released for the
entirety of the function.