Preserve array's order when raveling in `minmax`
By default `ravel` enforces `C`-order on any array it unravels. However
if the array is `F`-order or some other kind of order, it will copy the
array to perform the unraveling. However the underlying `cminmax`
function is unconcerned with the order and is only concerned that it is
1-D. So relax `ravel`'s behavior by allowing any known ordering through.
This includes `C`-order, `F`-order, and various non-traditional
orderings as well (e.g. reversing dimensions' directions, or swapping
axes). Should avoid unneeded copies in a few more cases, which will be a
significant help to `minmax`'s performance.