Simplify the `bool`-case of `minmax`
Instead of trying to create NumPy `uint8` views onto the NumPy `bool`
arrays, simply allow the type selection of the `cminmax` function and
the typed `memoryviews` to handle that themselves. This seems to be
sufficient to get the right behavior and drops the overhead associated
with creating two new Python objects.