Drop Python 3.5 on AppVeyor
Since the current matrix used here for AppVeyor is taking a bit to run,
looking for ways to cutdown on the matrix. Cannot really drop Python 2.7
as VS 2008 is used there and falling out of compatibility with it would
be quite painful if we want to get back in sync. Further Python 3.6 is
necessary as it is the newest Python version available and uses VS 2015
(so UCRT). Given Python 3.5 and 3.6 are not that different and both use
VS 2015 and UCRT, drop the older of the two (Python 3.5). This should
lighten up AppVeyor a bit and help it finish more quickly.