Use NumPy's type enum in `if` statements
Instead of comparing NumPy dtypes directly when determining which
`cminmax` type specialization to call, simply compare NumPy's type enum
values. This simplifies the checks as they are now just comparisons of
integral values where the left hand side is always the same. Given this
structure, Cython will try to optimize this down to a `switch` of `case`
statements. Further optimizations are therefore possible on the compiler
end to allow faster traversal through the type dispatch cases.