Allocate C variables for max and min on the stack
Instead of trying to store the max and min values directly into `out`
during each iteration, store the max and min values to two C variables
allocated on the stack at the start of this function. Cuts down on
overhead that is coupled with storing to `out` directly. This makes for
a notable speedup in the `for`-loops performance and thus the `minmax`
function's performance.