Operate on C types directly in `cminmax`
To avoid more unnecessary checks related to `memoryview`s inside the
`for`-loop, switch to working on C pointers within the `memoryview`s to
bypass these checks. Use the C pointers both to navigate through the
input array and to write results directly to the output array. The
resulting C code that Cython generates for this is pretty close to what
hand written C code outside of a Python context would look like. Thus
should be getting pretty close to optimal C-level performance.