Drop `binding` directive from `cminmax`
As Cython's default is already to set `binding` to `False`, there is not
much value in forcing `binding` to `False` for `cminmax`. Further it
seems that using this directive on `cminmax` overrides any global
directive for `binding`. As a result if `binding` is set to `True` at
the global level to allow profiling, this blocks our ability to profile
`cminmax` unless we remove the directive from the function as well.
Hence we go ahead and drop the `binding` directive from `cminmax` to
make profiling easier.