Use `arr_ptr` for initial min and max
Missed switching from `arr` to `arr_ptr` when setting the initial min
and max values when making the move from using the `memoryview` to using
a C pointer to the data within the `memoryview`. This fixes that error
by ensuring the C pointer is used to set the initial min and max values
as well.