Optimize comparisons in `cminmax`
Cutdown the number of comparisons done in `cminmax` by roughly one
quarter. This is done by tracking two neighboring positions and
comparing the values of those two positions. From this information, the
larger of the two values can be compared to the max only and the smaller
of the two values can be compared to the min only. This results in
shaving off on extra comparison that either of the two values would go
through.