Ensure `arr` owns its data
In order to ensure the data pointer of the array doesn't have any weird
offset for the first values or any other values, make sure that the
array owns its data. If the array doesn't own its data, make a copy of
the array so the copy will own its data.