Use NumPy's data pointer directly
Instead of using a `memoryview` to get at the data inside the input
array, simply use the data's array pointer. Also pass along the array's
size as well. This seems to result in a small performance boost. Not to
mention this avoids any need for things like `ravel`, which are a bit
more general than necessary.