Use `PyArray_CHKFLAGS` to test data ownership
Accessing the `flags` attribute of `arr` and then the `owndata`
attribute of `flags` is a bit slow due to the Python GIL and a bunch of
unnecessary general code. To do this a bit faster simply use the
`PyArray_CHKFLAGS` macro. This amounts to a struct member retrieval and
a mask operation. Both are very fast, occur in C, and thus do not
require the GIL.